requirements.txt
flask
flask-cors
flask_sqlalchemy
bcrypt
streamlit
requests
PyJWT
pymysql

accelerate
bitsandbytes
langchain
langchain-community
sentence-transformers
openpyxl
pacmap
pypdf
faiss-cpu
langchain-huggingface
langchain-text-splitters
langchain_experimental
langchain-community
sentence_transformers
chromadb
langchain-chroma
